+++ backend/main.py
@@ -1,1 +1,4 @@
-../app/main.py
+from app import create_app
+
+if __name__ == '__main__':
+    create_app().run(host='0.0.0.0', port=4040, threaded=True)
