+++ backend/scrapers/party_seed.py
